Figuring out roots is hard
Handle arbitrary directory separators.

Allow specifying the target via switch (to handle targets that start with the switch character).

Change ReplaceRoot to be a little simpler. Ideally we could make a two stage pass in the case where the root changes to find the _real_ casing for the Windows objects. (And possibly even a third pass for unc type paths to handle server/share)